requirements.txt
Django
psycopg2
dj_database_url
selenium
msedge-selenium-tools